fix: 优化多个服务中的查询逻辑，简化计数查询并移除不必要的结构体

+++ src/models/order.rs
@@ -29,7 +29,7 @@ impl From<order_entity::Model> for OrderResponse {
             product_name: m.product_name,
             price: m.price,
             stamps_earned: m.stamps_earned.unwrap_or(0),
-            order_status: m.order_status as i32,
+            order_status: m.order_status,
             external_created_at: m.external_created_at,
         }
     }
